fix(InpHelper): use minmax to fill interior of coreshape with 1
More pleasing calculated coreshape, filling in the interior based
on the display technique from ImageGenerator.
